requirements.txt
requests
pandas 
numpy
matplotlib
pytz
python-dotenv